Adds generic sensors and generic observers
Uses SensorStateMachines and SystemStateMachines to hold states and state transition logic. A StateManager is notified of any sensor driven state changes and tells the system dependent state machines to reassess their state after checking the relative upstream state machines.

+++ src/sensors/specific_sensors/test_i2c_sensor.h
@@ -4,14 +4,14 @@
 #include <src/sensors/generic_sensor.h>
 #include <string>
 
-class TestI2cSensor : public GenericSensor<int> {
+class TestI2cSensor : public GenericSensor<double> {
    public:
     explicit TestI2cSensor(std::string id);
     void TakeReading();
-    void SetDummySensorData(int dummy_data);
+    void SetDummySensorData(double dummy_data);
 
    private:
-    int dummy_data;
+    double dummy_data;
     std::string id;
 };
